+++ packages/expo-router/CHANGELOG.md
@@ -29,6 +29,8 @@
 - fix(web): fitToContents modal support on desktop, aligned default styles more with iOS ([#38028](https://github.com/expo/expo/pull/38028) by [@hirbod](https://github.com/hirbod))
 - fix navigation from preview to modal ([#37832](https://github.com/expo/expo/pull/37832) by [@Ubax](https://github.com/Ubax))
 - Fix children update in modal ([#38064](https://github.com/expo/expo/pull/38064) by [@Ubax](https://github.com/Ubax))
+- use portal in Modal ([#37980](https://github.com/expo/expo/pull/37980) by [@Ubax](https://github.com/Ubax))
+
 
 ### 💡 Others